- 3rdparty cleanup: Removed unused misleading zlib stuff and updated liborigin from its SVN repos - improved README.html (license section) and renamed it to INSTALL.html which IHMO fits better - added a plain text README containing the the most important information about SciDAVis - added a CHANGES file to contain future changelogs - adjusted scidavis.pro accordingly

+++ 3rdparty/liborigin/OPJFile.h
@@ -166,10 +166,10 @@ struct graphCurve {
 	int fillarea_pattern;
 	int fillarea_color;
 	int fillarea_pattern_color;
-	int fillarea_pattern_width;
+	double fillarea_pattern_width;
 	int fillarea_pattern_border_style;
 	int fillarea_pattern_border_color;
-	int fillarea_pattern_border_width;
+	double fillarea_pattern_border_width;
 
 
 	int symbol_type;
